inventory_optimization_bundle/pyproject.toml
[project]
name = "2026-buildathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash==3.3.0",
    "pandas==2.3.3",
    "databricks-sql-connector==4.2.4",
    "databricks-sdk==0.77.0",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "blinker==1.9.0",
    "cachelib==0.13.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorcet==3.1.0",
    "dash-ag-grid==32.3.4",
    "dash-bootstrap-components==2.0.4",
    "dash-extensions==2.0.4",
    "dash-leaflet==1.1.3",
    "dash-svg==0.0.12",
    "dataclass-wizard==0.39.1",
    "distro==1.9.0",
    "editorconfig==0.17.1",
    "et-xmlfile==2.0.0",
    "flask==3.1.2",
    "flask-caching==2.3.1",
    "geobuf==2.0.0",
    "google-auth==2.47.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jsbeautifier==1.15.4",
    "lz4==4.4.5",
    "markupsafe==3.0.3",
    "more-itertools==10.8.0",
    "narwhals==2.15.0",
    "nest-asyncio==1.6.0",
    "numpy==2.4.1",
    "oauthlib==3.3.1",
    "openai==2.15.0",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "plotly==6.5.1",
    "protobuf==6.33.4",
    "pyarrow==22.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybreaker==1.4.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "retrying==1.4.2",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "thrift==0.20.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "werkzeug==3.1.5",
    "zipp==3.23.0",
]
